requirements.txt
requests
PyPDF2
beautifulsoup4
python-dotenv
selenium

google-api-python-client
google-auth-httplib2
google-auth-oauthlib


